Extract resolve-relevant envs together
