+++ dhall/src/semantics/resolve/mod.rs
@@ -1,6 +1,6 @@
 pub mod env;
-pub mod resolve;
 pub mod hir;
+pub mod resolve;
 pub(crate) use env::*;
-pub(crate) use resolve::*;
 pub(crate) use hir::*;
+pub(crate) use resolve::*;
